grouped compound declarations in a directory
