+++ pkgs/swiftgen/swift2objc/lib/src/ast/_core/interfaces/confomingable.dart
@@ -1,4 +1,4 @@
-import '../../declarations/protocol_declaration.dart';
+import '../../declarations/compounds/protocol_declaration.dart';
 
 abstract interface class Conformingable {
   abstract List<ProtocolDeclaration> conformedProtocols;
