+++ pkgs/swiftgen/swift2objc/lib/src/ast/declarations/compounds/protocol_declaration.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../_core/interfaces/compund_declaration.dart';
-import '../_core/shared/parameter.dart';
-import '../_core/shared/referred_type.dart';
+import '../../_core/interfaces/compund_declaration.dart';
+import '../../_core/shared/parameter.dart';
+import '../../_core/shared/referred_type.dart';
 
 class ProtocolDeclaration implements CompoundDeclaration {
   @override
